实现了串口接收中断传输、DMA传输配合闲时中断；实现了CMD指令解析与执行功能

+++ MasterChip/CODE/motor.c
@@ -14,26 +14,25 @@
 #include "sci_compute.h"
 #include "pid.h"
 
-void
-Motor_Init (void)
+void Motor_Init(void)
 {
-  //初始化电机PWM引脚和方向引脚
+	//初始化电机PWM引脚和方向引脚
 
-  //桌大大的推文中，建议电磁组电机频率选用13K-17K
-  //最大占空比值PWM_DUTY_MAX 可以在zf_pwm.h文件中修改 默认为50000
-  //对于一个PWM模块 包含的所有通道只能输出频率一样 占空比不一样的 PWM CH32V103R8T6只有四个PWM模块 每个模块有4个通道
-  gpio_init (MOTOR1_A, GPO, 0, GPIO_PIN_CONFIG);
-  pwm_init (MOTOR1_B, 17000, 0);
-  gpio_init (MOTOR2_A, GPO, 0, GPIO_PIN_CONFIG);
-  pwm_init (MOTOR2_B, 17000, 0);
-  gpio_init (MOTOR3_A, GPO, 0, GPIO_PIN_CONFIG);
-  pwm_init (MOTOR3_B, 17000, 0);
-  gpio_init (MOTOR4_A, GPO, 0, GPIO_PIN_CONFIG);
-  pwm_init (MOTOR4_B, 17000, 0);
+	//桌大大的推文中，建议电磁组电机频率选用13K-17K
+	//最大占空比值PWM_DUTY_MAX 可以在zf_pwm.h文件中修改 默认为50000
+	//对于一个PWM模块 包含的所有通道只能输出频率一样 占空比不一样的 PWM CH32V103R8T6只有四个PWM模块 每个模块有4个通道
+	gpio_init(MOTOR1_A, GPO, 0, GPIO_PIN_CONFIG);
+	pwm_init(MOTOR1_B, 17000, 0);
+	gpio_init(MOTOR2_A, GPO, 0, GPIO_PIN_CONFIG);
+	pwm_init(MOTOR2_B, 17000, 0);
+	gpio_init(MOTOR3_A, GPO, 0, GPIO_PIN_CONFIG);
+	pwm_init(MOTOR3_B, 17000, 0);
+	gpio_init(MOTOR4_A, GPO, 0, GPIO_PIN_CONFIG);
+	pwm_init(MOTOR4_B, 17000, 0);
 }
 // kp ki kd int_duty int_max int_sum last_err last_delta_err
 PID_t MotorPID =
-  { 0.1, 0.01, 0.0, 10, 10, 0, 0, 0 };
+{ 0.1, 0.01, 0.0, 10, 10, 0, 0, 0 };
 /**
  * @brief 电机占空比环
  * 
@@ -41,8 +40,7 @@ PID_t MotorPID =
  * @param now_rpm 
  * @param expect_speed 
  */
-void
-Motor_DutyCtrl (int16 target_duty)
+void Motor_DutyCtrl(int16 target_duty)
 {
 //    int16 pwm_out = PID_GetOutput(&MotorPID,target_duty);
 //    return pwm_out;
@@ -55,55 +53,54 @@ Motor_DutyCtrl (int16 target_duty)
  * @param
  * @note
  */
-void
-Motor_SetDuty (int32 duty_1, int32 duty_2, int32 duty_3, int32 duty_4)
+void Motor_SetDuty(int32 duty_1, int32 duty_2, int32 duty_3, int32 duty_4)
 {
-  //对占空比限幅
-  LIMIT(duty_1, PWM_DUTY_MAX);
-  LIMIT(duty_2, PWM_DUTY_MAX);
-  LIMIT(duty_3, PWM_DUTY_MAX);
-  LIMIT(duty_4, PWM_DUTY_MAX);
+	//对占空比限幅
+	LIMIT(duty_1, PWM_DUTY_MAX);
+	LIMIT(duty_2, PWM_DUTY_MAX);
+	LIMIT(duty_3, PWM_DUTY_MAX);
+	LIMIT(duty_4, PWM_DUTY_MAX);
 
-  // 设置占空比为 百分之 (1000/TIMER1_PWM_DUTY_MAX*100)
-  if (0 <= duty_1) //电机1   正转
-    {
-      gpio_set (MOTOR1_A, FORWARD_ROTATE);
-    }
-  else //电机1   反转
-    {
-      gpio_set (MOTOR1_A, REVERSE_ROTATE);
-    }
-  pwm_duty (MOTOR1_B, duty_1);
+	// 设置占空比为 百分之 (1000/TIMER1_PWM_DUTY_MAX*100)
+	if (0 <= duty_1) //电机1   正转
+	{
+		gpio_set(MOTOR1_A, FORWARD_ROTATE);
+	}
+	else //电机1   反转
+	{
+		gpio_set(MOTOR1_A, REVERSE_ROTATE);
+	}
+	pwm_duty(MOTOR1_B, duty_1);
 
-  if (0 <= duty_2) //电机2   正转
-    {
-      gpio_set (MOTOR2_A, FORWARD_ROTATE);
-    }
-  else //电机2   反转
-    {
-      gpio_set (MOTOR2_A, REVERSE_ROTATE);
-    }
-  pwm_duty (MOTOR2_B, duty_2);
+	if (0 <= duty_2) //电机2   正转
+	{
+		gpio_set(MOTOR2_A, FORWARD_ROTATE);
+	}
+	else //电机2   反转
+	{
+		gpio_set(MOTOR2_A, REVERSE_ROTATE);
+	}
+	pwm_duty(MOTOR2_B, duty_2);
 
-  if (0 <= duty_3) //电机3   正转
-    {
-      gpio_set (MOTOR3_A, FORWARD_ROTATE);
-    }
-  else //电机3   反转
-    {
-      gpio_set (MOTOR3_A, REVERSE_ROTATE);
-    }
-  pwm_duty (MOTOR3_B, duty_3);
+	if (0 <= duty_3) //电机3   正转
+	{
+		gpio_set(MOTOR3_A, FORWARD_ROTATE);
+	}
+	else //电机3   反转
+	{
+		gpio_set(MOTOR3_A, REVERSE_ROTATE);
+	}
+	pwm_duty(MOTOR3_B, duty_3);
 
-  if (0 <= duty_4) //电机3   正转
-    {
-      gpio_set (MOTOR4_A, FORWARD_ROTATE);
-    }
-  else //电机3   反转
-    {
-      gpio_set (MOTOR4_A, REVERSE_ROTATE);
-    }
-  pwm_duty (MOTOR4_B, duty_4);
+	if (0 <= duty_4) //电机3   正转
+	{
+		gpio_set(MOTOR4_A, FORWARD_ROTATE);
+	}
+	else //电机3   反转
+	{
+		gpio_set(MOTOR4_A, REVERSE_ROTATE);
+	}
+	pwm_duty(MOTOR4_B, duty_4);
 }
 
 /**
@@ -112,16 +109,15 @@ Motor_SetDuty (int32 duty_1, int32 duty_2, int32 duty_3, int32 duty_4)
  * @param speed m/s
  * @return
  */
-float
-SW_Speed2PWM (float speed)
+int SW_Speed2PWM(float speed)
 {
+
 }
 
-void
-DriveMotors_LimitSpeed (float speed[4])
+void DriveMotors_LimitSpeed(float speed[4])
 {
-  for (int i = 0; i < 4; i++)
-    {
-      LIMIT_FROM_TO(speed[i], DRIVE_WHEEL_MIN_SPEED, DRIVE_WHEEL_MAX_SPEED);
-    }
+	for (int i = 0; i < 4; i++)
+	{
+		LIMIT_FROM_TO(speed[i], DRIVE_WHEEL_MIN_SPEED, DRIVE_WHEEL_MAX_SPEED);
+	}
 }
